requirements.txt
streamlit
ultralytics
torch
torchvision
pillow
pandas
opencv-python-headless==4.8.1.78